Fix type of aod-timeframe-id to uint64
Internally uint64 was already used but inconsistent
with the command line option.

Now it should be ok to use ALIEN_PROC_ID + xxx as identification.

+++ Detectors/AOD/include/AODProducerWorkflow/AODProducerWorkflowSpec.h
@@ -136,7 +136,7 @@ class AODProducerWorkflowDPL : public Task
   int mFillTracksITS{1};
   int mFillTracksTPC{0};
   int mFillTracksITSTPC{1};
-  int mTFNumber{-1};
+  uint64_t mTFNumber{-1};
   int mTruncate{1};
   int mIgnoreWriter{0};
   int mRecoOnly{0};
